added compatability for mac

+++ install.sh
@@ -1,6 +1,7 @@
 #!/usr/bin/env bash
 
 unameOut="$(uname -s)"
+_uname="$(logname)"
 case "${unameOut}" in
     Linux*)     machine=Linux;;
     Darwin*)    machine=Mac;;
@@ -11,7 +12,7 @@ esac
 
 # configuration
 if [[ "$machine" = "Mac" ]]; then
-    readonly attendance_folder_path="/Users/$(logname)/.attendance"
+    readonly attendance_folder_path="/Users/$_uname/.attendance"
 else
     readonly attendance_folder_path="/opt/attendance"
     readonly labtrac_service_path="/etc/systemd/system"
@@ -25,19 +26,26 @@ git clone https://github.com/amfoss/attendance-tracker.git
 sudo mkdir -p "$attendance_folder_path"
 # remove all old contents if any
 sudo rm -rf "$attendance_folder_path"/*
-sudo rm -f "$labtrac_service_path"/labtrac.service
-sudo rm -f "$labtrac_service_path"/labtrac.timer
+if [[ "$machine" != "Mac" ]]; then
+    sudo rm -f "$labtrac_service_path"/labtrac.service
+    sudo rm -f "$labtrac_service_path"/labtrac.timer
+fi
 
 sudo cp -r attendance-tracker/attendance/. "$attendance_folder_path"/.
 
 sudo chmod +x "$attendance_folder_path"/config "$attendance_folder_path"/get_ssid_names.sh
 
-sudo cp -r attendance-tracker/system/. "$labtrac_service_path"/.
-
 # Activate the service
-sudo systemctl enable labtrac.timer
-sudo systemctl start labtrac.service
+if [[ "$machine" != "Mac" ]]; then
+    sudo cp -r attendance-tracker/system/. "$labtrac_service_path"/.
+    sudo systemctl enable labtrac.timer
+    sudo systemctl start labtrac.service
+fi
 
+if [[ "$machine" = "Mac" ]]; then
+    sudo chmod u+x macinstall.sh
+    sudo ./macinstall.sh $_uname
+fi
 # delete downloaded files
 rm -rf attendance-tracker
 rm install.sh
